requirements.txt
langchain-google-genai
langchain-neo4j
graphdatascience
python-dotenv
langchain-openai
langchain-core
langchain-community
pillow